# docs/swagger.yaml
definitions:
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_student_Student:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_student.Student'
        type: array
      message:
        type: string
    type: object
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_student_ImportResult:
    properties:
      data:
        $ref: '#/definitions/internal_student.ImportResult'
      message:
        type: string
    type: object
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_student_Student:
    properties:
      data:
        $ref: '#/definitions/internal_student.Student'
      message:
        type: string
    type: object
...
  github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string:
    properties:
      data:
        $ref: '#/definitions/map_string_string'
      message:
        type: string
    type: object
...
  internal_student.ImportResult:
    properties:
      enrollmentsAdded:
        type: integer
      errors:
        items:
          type: string
        type: array
      inserted:
        type: integer
      updated:
        type: integer
    type: object
  internal_student.Student:
    properties:
      annotation:
        type: string
      consent:
        type: boolean
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        $ref: '#/definitions/internal_student.StudentStatus'
      studentId:
        type: string
    type: object
  internal_student.StudentStatus:
    enum:
    - ACTIVE
    - CANCELED
    - GRADUATED
    - LOCKED
    - PENDING
    type: string
    x-enum-varnames:
    - StudentStatusActive
    - StudentStatusCanceled
    - StudentStatusGraduated
    - StudentStatusLocked
    - StudentStatusPending
  internal_student.createStudentInput:
    properties:
      annotation:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/internal_student.StudentStatus'
        enum:
        - ACTIVE
        - CANCELED
        - GRADUATED
        - LOCKED
      studentId:
        type: string
    required:
    - status
    - studentId
    type: object
  internal_user.createUserInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  map_string_string:
    additionalProperties:
      type: string
    type: object
info:
  contact: {}
paths:
...
  /course/{courseId}/students/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: upsert ou clean
        enum:
        - upsert
        - clean
        in: query
        name: mode
        type: string
      - description: CSV com studentId,name,phone,email,status
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_student_ImportResult'
      summary: Importa estudantes para um curso (CSV)
      tags:
      - student
...
  /student:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Program ID
        in: query
        name: program
        type: string
      - description: Campus ID
        in: query
        name: campus
        type: string
      - description: Course ID
        in: query
        name: course
        type: string
      - description: User ID
        in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-array_internal_student_Student'
      summary: Lista estudantes com filtros
      tags:
      - student
  /student/{id}:
    delete:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Remove um estudante
      tags:
      - student
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-internal_student_Student'
      summary: Obtém um estudante por ID
      tags:
      - student
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Student ID
        in: path
        name: id
        required: true
        type: string
      - description: Campos para atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_student.createStudentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Atualiza um estudante
      tags:
      - student
  /student/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados do estudante
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_student.createStudentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria um estudante (pré-cadastro)
      tags:
      - student
  /user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dados do usuário
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_user.createUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ThalysSilva_unicast-backend_pkg_api.DefaultResponse-map_string_string'
      summary: Cria um usuário
      tags:
      - user
...
swagger: "2.0"